refactor: extract stripDiacritics helper to mongoUtils
