+++ server/src/common/utils/mongoUtils.ts
@@ -1,6 +1,57 @@
 import { WithId } from "mongodb";
 
 /**
+ * Normalisation des accents en agrégation Mongo
+ */
+export function stripDiacritics(expr: Record<string, unknown>) {
+  const replacements = [
+    ["é", "e"],
+    ["è", "e"],
+    ["ê", "e"],
+    ["ë", "e"],
+    ["à", "a"],
+    ["â", "a"],
+    ["ä", "a"],
+    ["ù", "u"],
+    ["û", "u"],
+    ["ü", "u"],
+    ["ô", "o"],
+    ["ö", "o"],
+    ["î", "i"],
+    ["ï", "i"],
+    ["ç", "c"],
+    ["ÿ", "y"],
+    ["É", "e"],
+    ["È", "e"],
+    ["Ê", "e"],
+    ["Ë", "e"],
+    ["À", "a"],
+    ["Â", "a"],
+    ["Ä", "a"],
+    ["Ù", "u"],
+    ["Û", "u"],
+    ["Ü", "u"],
+    ["Ô", "o"],
+    ["Ö", "o"],
+    ["Î", "i"],
+    ["Ï", "i"],
+    ["Ç", "c"],
+    ["Ÿ", "y"],
+    ["́", ""],
+    ["̀", ""],
+    ["̂", ""],
+    ["̈", ""],
+    ["̧", ""],
+  ];
+
+  let result: Record<string, unknown> = expr;
+  for (const [from, to] of replacements) {
+    result = { $replaceAll: { input: result, find: from, replacement: to } };
+  }
+  return result;
+}
+
+/**
  * Supprime toutes les attributs à false ou 0 pour éviter une erreur mongodb
  */
 export function cleanProjection<
